Sentiment Analysis with Bert model
In this study, in line with the distance education decision announced on 11 February, I conducted a sentiment analysis of 36,910 tweets posted on Twitter with the keyword "university".
After obtaining the tweets between 11.02.2023-20.02.2023 with the necessary library, unnecessary expressions were deleted from the tweets.
The BERT model, which provides high achievements in the field of natural language processing and is frequently used today to classify Turkish texts, was preferred.

As a result of the use of the pre-trained model, the emotional states were visualised as a percentage and the emotional state in the following days was observed.
In addition, the keywords frequently used in the tweets were visualised with the Word Cloud library.
The study was carried out in order to apply and reinforce what was learnt and was concluded without any interpretation.